% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{add_distance}
\alias{add_distance}
\title{Add distance}
\usage{
add_distance(track, crs = NULL)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{crs}{object compatible with sf::st_crs -- will be derived from gpxr::get_utm}
}
\description{
Add distance
}
