% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{clean_elev}
\alias{clean_elev}
\title{Clean Elevation}
\usage{
clean_elev(track, window = 4)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{window}{integer number of points to average over.}
}
\description{
executes a sliding mean over the specified window.
}
