% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{cut_track}
\alias{cut_track}
\title{cut track}
\usage{
cut_track(track, cut_one, cut_two)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{cut_one}{integer track_seg_point_id for start of cut}

\item{cut_two}{integer track_seg_point_id for end of cut}
}
\description{
cuts a gpx track at two specified points
}
