% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{load_track_points}
\alias{load_track_points}
\title{load track points}
\usage{
load_track_points(track_f)
}
\arguments{
\item{track_f}{character path to track gpx file}
}
\description{
load track points
}
