% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{make_loop}
\alias{make_loop}
\title{make loop}
\usage{
make_loop(
  track,
  start_id,
  end_id,
  lap_start = NULL,
  control = NULL,
  correct_elevation = FALSE
)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{start_id}{numeric track_seg_point_id of start}

\item{end_id}{numeric track_seg_point_id of end}

\item{lap_start}{track_seg_point_id of desired start --
NULL for no adjustment.}

\item{control}{numeric c(lon, lat)}

\item{correct_elevation}{logical should elevation of start be forced
to match elevation of end?}
}
\description{
make loop
}
