% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{resample_track}
\alias{resample_track}
\title{resample_track}
\usage{
resample_track(track, l)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{l}{numeric distance between points}
}
\description{
resample track to a specific distance spacing
}
