% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{segmentize_track}
\alias{segmentize_track}
\title{segmentize track}
\usage{
segmentize_track(track, l)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{l}{numeric maximum segment length in units of data projection}
}
\description{
break track into segments
}
