% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{simplify_track}
\alias{simplify_track}
\title{simplify track}
\usage{
simplify_track(track, tolerance)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{tolerance}{numeric passed to dTolerance}
}
\description{
convert to line and simplify
}
