% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{write_track_points}
\alias{write_track_points}
\title{write track points}
\usage{
write_track_points(track, track_f)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{track_f}{character output track file}
}
\description{
write track points
}
