% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{bez_smooth}
\alias{bez_smooth}
\title{bezier smooth}
\usage{
bez_smooth(track, start_id, end_id, control)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{start_id}{numeric track_seg_point_id of start}

\item{end_id}{numeric track_seg_point_id of end}

\item{control}{numeric c(lon, lat)}
}
\description{
bezier smooth
}
