% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_edit.R
\name{buffer_track}
\alias{buffer_track}
\title{buffer track}
\usage{
buffer_track(track, proj = 5070)
}
\arguments{
\item{track}{sf data.frame gpx track}

\item{proj}{input to sf::st_crs to buffer points in}
}
\description{
buffer track
}
